calloc malloc(0) etc..

+++ ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: hohnuki <hohnuki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/11 14:15:12 by hohnuki           #+#    #+#             */
-/*   Updated: 2021/10/29 19:04:24 by hohnuki          ###   ########.fr       */
+/*   Updated: 2021/11/08 14:58:36 by hohnuki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strmapi(const char *s, char (*f)(unsigned int, char))
 {
-	char			*str;
-	unsigned int	i;
+	char	*str;
+	size_t	i;
 
 	if (!s)
 		return (NULL);
